relocate getRadioSchedulesUrls
