+++ src/app/containers/RadioSchedule/index.jsx
@@ -7,7 +7,7 @@ import { RequestContext } from '#contexts/RequestContext';
 import useToggle from '#hooks/useToggle';
 import Canonical from './Canonical';
 import radioSchedulesShape from './utilities/radioScheduleShape';
-import { getRadioScheduleEndpoint } from '#lib/utilities/getRadioSchedulesUrls';
+import { getRadioScheduleEndpoint } from '#lib/utilities/getUrlHelpers';
 
 const RadioScheduleContainer = ({
   initialData,
